Add files via upload
In this Problem I have sloved 8 Queen puzzle by using Hill CLimb Algorithm. The problem with hill climb is that it can stuck at 14% success rate. When the local maxima is found and we  have ridges , the  the algo move sideways up to a certain number of iterations. By doing this it can escape the local maxima and global maxima can be achieved and success rate goes from 14% to 94% . Just remember that the success comes at a cost.